create new supertype InteractionRuntimeExecutable for DynamicOptionResolverDefinition
